Fix -Wc++11-narrowing: content/
Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2975065
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#894183}

+++ content/app_shim_remote_cocoa/render_widget_host_ns_view_bridge.mm
@@ -255,7 +255,7 @@ void RenderWidgetHostNSViewBridge::ShowDictionaryOverlay(
   if ([string length] == 0)
     return;
   NSPoint flipped_baseline_point = {
-      baseline_point.x(),
+      static_cast<CGFloat>(baseline_point.x()),
       [cocoa_view_ frame].size.height - baseline_point.y(),
   };
   [cocoa_view_ showDefinitionForAttributedString:string
